requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncclick==8.1.8.0
certifi==2025.7.14
charset-normalizer==3.4.2
dictdiffer==0.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypika-tortoise==0.6.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-telegram-bot==22.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.12.3
sniffio==1.3.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
vcrpy==7.0.0
wrapt==1.17.2
yarl==1.20.1
